requirements.txt
streamlit>=1.36.0
langchain>=0.2.0
langchain-google-genai>=1.0.1
langchain-huggingface>=0.0.1
langchain-community>=0.2.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=2.6.1
PyMuPDF>=1.24.1
pdfplumber>=0.11.0
pandas>=2.2.1
plotly>=5.20.0
pydantic>=2.6.4
python-dotenv>=1.0.1
fpdf2>=2.7.8
langchain-groq>=0.1.3
langchain-ollama>=0.0.1
tenacity>=9.1.4
